Initial work in moving hooks to their own files.
