+++ src/CMakeLists.txt
@@ -14,7 +14,7 @@ set(GAMEENGINE_INCLUDES
     game/util
 )
 
-set(HOOKER_SRC hooker/dllmain.cpp hooker/hooker.cpp)
+set(HOOKER_SRC hooker/dllmain.cpp hooker/hooker.cpp hooker/setupglobals.cpp hooker/setuphooks.cpp)
 
 set(GAMEENGINE_SRC
     game/main.cpp
