+++ src/hooker/setupglobals.cpp
@@ -1,0 +1,1 @@
+#include "hooker.h"
