+++ src/hooker/setuphooks.h
@@ -1,0 +1,21 @@
+/**
+ * @file
+ *
+ * @author OmniBlade
+ *
+ * @brief Function to hook all required functions.
+ *
+ * @copyright Chronoshift is free software: you can redistribute it and/or
+ *            modify it under the terms of the GNU General Public License
+ *            as published by the Free Software Foundation, either version
+ *            2 of the License, or (at your option) any later version.
+ *            A full copy of the GNU General Public License can be found in
+ *            LICENSE
+ */
+
+#ifndef SETUP_HOOKS_H
+#define SETUP_HOOKS_H
+
+void Setup_Hooks();
+
+#endif
